备份下

+++ src/main/java/com/alibaba/middleware/handlefile/FileIndexWithOffset.java
@@ -7,10 +7,10 @@ package com.alibaba.middleware.handlefile;
  */
 public class FileIndexWithOffset {
 
-	public int fileIndex = 0;
+	public short fileIndex = 0;
 	public long offset = 0;
 	
-	public FileIndexWithOffset( int fileIndex, long offset) {
+	public FileIndexWithOffset( short fileIndex, long offset) {
 		this.fileIndex = fileIndex;
 		this.offset = offset;
 	}
